Use UUID instead of the device name in fstab and GRUB configuration
Add the mapping from UUID to device for the root file system to the
chroot. This allows update-grub to generate the root= linux command
line parameter that uses the UUID instead of the device name.

Also use the UUID for the EFI partition when generating /etc/fstab.

This should make the installed system independent of the name of the
root disk.