Support multiple postRootFs command hooks
Rename the postRootFs{Create,Install}Cmd arguments to
postRootFs{Create,Install}Cmds and change the semantics to lists of
commands instead of a single command.

Also copy the closures of all commands to the Nix store in the chroot
to avoid missing dependencies when the commands are executed in the
chroot.