install.sh: use dedicated mount point for chroot, mount efivarfs
Set NOS mode with /sbin/onie-nos-mode -s in the installer after
successful completion to conform to the ONIE requirements. Using /mnt
as chroot mount point conflicts with this, so we use mktemp to create
a unique mount point instead. Also mount efivarfs in the chroot, used
by grub-install. This appears to fix a problem on certain targets
where the boot entries generated with efibootmgr are lost after a
reboot.

Perform proper unmounts for all file systems mounted by the installer.

Let the ONIE installer perform the reboot instead of forcing it early
in install.sh.