install.sh: serialize umounts, don't set onie-nos-mode
It appears that the busybox umount either doesn't respect the order of
file systems given on the command line or has a race-condition to that
effect, which can cause the umount to fail. Serializing the umounts
should solve this problem.

Setting onie-nos-mode has the side-effect of putting ONIE in rescue
mode, which is counter-intuitive if ONIE is entered via the
chainloader from the NOS boot loader and could prevent automated
re-installs from working. This effect could be avoided by setting the
onie_mode Grub environment variable as well, but this seems
complicated and not setting onie-nos-mode appears to have no bad
effects either, so we drop it.